+++ smartcontract/sdk/rs/src/commands/feed/delete.rs
@@ -1,4 +1,4 @@
-use crate::DoubleZeroClient;
+use crate::{commands::common::append_payer_permission_account, DoubleZeroClient};
 use doublezero_serviceability::processors::feed::delete::FeedDeleteArgs;
 use doublezero_serviceability_instruction::feed::delete_feed;
 use solana_sdk::{pubkey::Pubkey, signature::Signature};
@@ -10,25 +10,33 @@ pub struct DeleteFeedCommand {
 
 impl DeleteFeedCommand {
     pub fn execute(&self, client: &dyn DoubleZeroClient) -> eyre::Result<Signature> {
-        client.send_transaction(delete_feed(
+        let mut ix = delete_feed(
             &client.get_program_id(),
             &client.get_payer(),
             &self.pubkey,
             FeedDeleteArgs {},
-        ))
+        );
+        append_payer_permission_account(client, &mut ix);
+        client.send_transaction(ix)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use crate::{
-        commands::feed::delete::DeleteFeedCommand, tests::utils::create_test_client,
+        commands::feed::delete::DeleteFeedCommand,
+        tests::utils::{create_test_client, expect_missing_permission_account},
         DoubleZeroClient,
     };
-    use doublezero_serviceability::{pda::get_feed_pda, processors::feed::delete::FeedDeleteArgs};
+    use doublezero_serviceability::{
+        pda::{get_feed_pda, get_permission_pda},
+        processors::feed::delete::FeedDeleteArgs,
+    };
     use doublezero_serviceability_instruction::feed::delete_feed;
     use mockall::predicate;
-    use solana_sdk::{pubkey::Pubkey, signature::Signature};
+    use solana_sdk::{
+        account::Account, message::AccountMeta, pubkey::Pubkey, signature::Signature,
+    };
 
     #[test]
     fn test_commands_feed_delete_command() {
@@ -43,6 +51,37 @@ mod tests {
             .expect_send_transaction()
             .with(predicate::eq(expected))
             .returning(|_| Ok(Signature::new_unique()));
+
+        expect_missing_permission_account(&mut client);
+
+        let res = DeleteFeedCommand { pubkey: pda_pubkey }.execute(&client);
+
+        assert!(res.is_ok());
+    }
+
+    #[test]
+    fn test_commands_feed_delete_command_with_permission_pda() {
+        let mut client = create_test_client();
+
+        let program_id = client.get_program_id();
+        let payer = client.get_payer();
+        let (pda_pubkey, _) = get_feed_pda(&program_id, "test_feed", &Pubkey::new_unique());
+
+        let mut expected = delete_feed(&program_id, &payer, &pda_pubkey, FeedDeleteArgs {});
+        let (permission_pda_pubkey, _) = get_permission_pda(&program_id, &payer);
+        expected
+            .accounts
+            .push(AccountMeta::new_readonly(permission_pda_pubkey, false));
+
+        client
+            .expect_send_transaction()
+            .with(predicate::eq(expected))
+            .returning(|_| Ok(Signature::new_unique()));
+
+        client
+            .expect_get_account()
+            .with(predicate::eq(permission_pda_pubkey))
+            .returning(move |_| Ok(Account::new(0, 0, &program_id)));
 
         let res = DeleteFeedCommand { pubkey: pda_pubkey }.execute(&client);